+++ pyoxidizer/src/starlark/python_interpreter_config.rs
@@ -4,12 +4,10 @@
 
 use crate::py_packaging::config::RunMode;
 use {
-    super::env::{get_context, EnvironmentContext},
-    super::util::{optional_list_arg, optional_str_arg, required_bool_arg, required_type_arg},
+    super::util::{optional_bool_arg, optional_int_arg, optional_list_arg, optional_str_arg},
     crate::py_packaging::config::{
         default_raw_allocator, EmbeddedPythonConfig, RawAllocator, TerminfoResolution,
     },
-    starlark::environment::TypeValues,
     starlark::values::error::{RuntimeError, ValueError, INCORRECT_PARAMETER_TYPE_ERROR_CODE},
     starlark::values::none::NoneType,
     starlark::values::{Immutable, TypedValue, Value, ValueResult},
@@ -40,7 +38,6 @@ impl TypedValue for EmbeddedPythonConfig {
 impl EmbeddedPythonConfig {
     /// Obtain a default instance from the context of Starlark.
     pub fn default_starlark() -> Self {
-        // Keep this in sync with default values from Starlark.
         Self {
             bytes_warning: 0,
             ignore_environment: true,
@@ -62,6 +59,7 @@ impl EmbeddedPythonConfig {
             sys_frozen: false,
             sys_meipass: false,
             sys_paths: Vec::new(),
+            // TODO this should come from the build context.
             raw_allocator: default_raw_allocator(crate::project_building::HOST),
             run_mode: RunMode::Repl,
             terminfo_resolution: TerminfoResolution::Dynamic,
@@ -74,7 +72,6 @@ impl EmbeddedPythonConfig {
     /// PythonInterpreterConfig(...)
     #[allow(clippy::too_many_arguments)]
     pub fn starlark_new(
-        type_values: &TypeValues,
         bytes_warning: &Value,
         ignore_environment: &Value,
         inspect: &Value,
@@ -106,46 +103,57 @@ impl EmbeddedPythonConfig {
         write_bytecode: &Value,
         write_modules_directory_env: &Value,
     ) -> ValueResult {
-        // TODO consider passing default of None for all arguments and using
-        // value from default_starlark() so logic doesn't get out of sync.
-        required_type_arg("bytes_warning", "int", &bytes_warning)?;
-        let ignore_environment = required_bool_arg("ignore_environment", &ignore_environment)?;
-        let inspect = required_bool_arg("inspect", &inspect)?;
-        let interactive = required_bool_arg("interactive", &interactive)?;
-        let isolated = required_bool_arg("isolated", &isolated)?;
+        let default = EmbeddedPythonConfig::default_starlark();
+
+        let bytes_warning = optional_int_arg("bytes_warning", &bytes_warning)?
+            .unwrap_or(default.bytes_warning as i64) as i32;
+        let ignore_environment = optional_bool_arg("ignore_environment", &ignore_environment)?
+            .unwrap_or(default.ignore_environment);
+        let inspect = optional_bool_arg("inspect", &inspect)?.unwrap_or(default.inspect);
+        let interactive =
+            optional_bool_arg("interactive", &interactive)?.unwrap_or(default.interactive);
+        let isolated = optional_bool_arg("isolated", &isolated)?.unwrap_or(default.isolated);
         let legacy_windows_fs_encoding =
-            required_bool_arg("legacy_windows_fs_encoding", &legacy_windows_fs_encoding)?;
+            optional_bool_arg("legacy_windows_fs_encoding", &legacy_windows_fs_encoding)?
+                .unwrap_or(default.legacy_windows_fs_encoding);
         let legacy_windows_stdio =
-            required_bool_arg("legacy_windows_stdio", &legacy_windows_stdio)?;
-        required_type_arg("optimize_level", "int", &optimize_level)?;
-        let parser_debug = required_bool_arg("parser_debug", &parser_debug)?;
+            optional_bool_arg("legacy_windows_stdio", &legacy_windows_stdio)?
+                .unwrap_or(default.legacy_windows_stdio);
+        let optimize_level =
+            optional_int_arg("optimize_level", &optimize_level)?.unwrap_or(default.optimize_level);
+        let parser_debug =
+            optional_bool_arg("parser_debug", &parser_debug)?.unwrap_or(default.parser_debug);
         let stdio_encoding = optional_str_arg("stdio_encoding", &stdio_encoding)?;
-        let unbuffered_stdio = required_bool_arg("unbuffered_stdio", &unbuffered_stdio)?;
-        let filesystem_importer = required_bool_arg("filesystem_importer", &filesystem_importer)?;
-        let quiet = required_bool_arg("quiet", &quiet)?;
+        let unbuffered_stdio = optional_bool_arg("unbuffered_stdio", &unbuffered_stdio)?
+            .unwrap_or(default.unbuffered_stdio);
+        let filesystem_importer = optional_bool_arg("filesystem_importer", &filesystem_importer)?
+            .unwrap_or(default.filesystem_importer);
+        let quiet = optional_bool_arg("quiet", &quiet)?.unwrap_or(default.quiet);
         let run_eval = optional_str_arg("run_eval", &run_eval)?;
         let run_file = optional_str_arg("run_file", &run_file)?;
         let run_module = optional_str_arg("run_module", &run_module)?;
-        let run_noop = required_bool_arg("run_noop", &run_noop)?;
-        let run_repl = required_bool_arg("run_repl", &run_repl)?;
-        let sys_frozen = required_bool_arg("sys_frozen", &sys_frozen)?;
-        let sys_meipass = required_bool_arg("sys_meipass", &sys_meipass)?;
+        let run_noop = optional_bool_arg("run_noop", &run_noop)?.unwrap_or(false);
+        let run_repl = optional_bool_arg("run_repl", &run_repl)?.unwrap_or(false);
+        let site_import =
+            optional_bool_arg("site_importer", &site_import)?.unwrap_or(default.site_import);
+        let sys_frozen =
+            optional_bool_arg("sys_frozen", &sys_frozen)?.unwrap_or(default.sys_frozen);
+        let sys_meipass =
+            optional_bool_arg("sys_meipass", &sys_meipass)?.unwrap_or(default.sys_meipass);
         optional_list_arg("sys_paths", "string", &sys_paths)?;
         let raw_allocator = optional_str_arg("raw_allocator", &raw_allocator)?;
-        let site_import = required_bool_arg("site_importer", &site_import)?;
         let terminfo_resolution = optional_str_arg("terminfo_resolution", &terminfo_resolution)?;
         let terminfo_dirs = optional_str_arg("terminfo_dirs", &terminfo_dirs)?;
-        let use_hash_seed = required_bool_arg("use_hash_seed", &use_hash_seed)?;
-        let user_site_directory = required_bool_arg("user_site_directory", &user_site_directory)?;
-        required_type_arg("verbose", "int", &verbose)?;
-        let write_bytecode = required_bool_arg("write_bytecode", &write_bytecode)?;
+        let use_hash_seed =
+            optional_bool_arg("use_hash_seed", &use_hash_seed)?.unwrap_or(default.use_hash_seed);
+        let user_site_directory = optional_bool_arg("user_site_directory", &user_site_directory)?
+            .unwrap_or(default.user_site_directory);
+        let verbose =
+            optional_int_arg("verbose", &verbose)?.unwrap_or(default.verbose as i64) as i32;
+        let write_bytecode =
+            optional_bool_arg("write_bytecode", &write_bytecode)?.unwrap_or(default.write_bytecode);
         let write_modules_directory_env =
             optional_str_arg("write_modules_directory_env", &write_modules_directory_env)?;
-
-        let raw_context = get_context(type_values)?;
-        let context = raw_context
-            .downcast_ref::<EnvironmentContext>()
-            .ok_or(ValueError::IncorrectParameterType)?;
 
         let mut run_count = 0;
         if run_eval.is_some() {
@@ -181,7 +189,7 @@ impl EmbeddedPythonConfig {
         } else if run_noop {
             RunMode::Noop
         } else {
-            RunMode::Repl
+            default.run_mode
         };
 
         let (stdio_encoding_name, stdio_encoding_errors) = if let Some(ref v) = stdio_encoding {
@@ -200,16 +208,17 @@ impl EmbeddedPythonConfig {
                     return Err(ValueError::from(RuntimeError {
                         code: INCORRECT_PARAMETER_TYPE_ERROR_CODE,
                         message: "invalid value for raw_allocator".to_string(),
-                        label: "invalid value for raw_allocator".to_string(),
+                        label: "PythonInterpreterConfig()".to_string(),
                     }));
                 }
             },
-            None => default_raw_allocator(&context.build_target_triple),
+            None => default.raw_allocator,
         };
 
         let terminfo_resolution = match terminfo_resolution {
             Some(x) => match x.as_ref() {
                 "dynamic" => TerminfoResolution::Dynamic,
+                "none" => TerminfoResolution::None,
                 "static" => TerminfoResolution::Static(if let Some(dirs) = terminfo_dirs {
                     dirs
                 } else {
@@ -217,18 +226,19 @@ impl EmbeddedPythonConfig {
                         code: INCORRECT_PARAMETER_TYPE_ERROR_CODE,
                         message: "terminfo_dirs must be set when using static resolution"
                             .to_string(),
-                        label: "terminfo_dirs must be set when using static resolution".to_string(),
+                        label: "PythonInterpreterConfig()".to_string(),
                     }));
                 }),
                 _ => {
                     return Err(ValueError::from(RuntimeError {
                         code: INCORRECT_PARAMETER_TYPE_ERROR_CODE,
-                        message: "terminfo_resolution must be 'dynamic' or 'static'".to_string(),
-                        label: "terminfo_resolution must be 'dynamic' or 'static'".to_string(),
+                        message: "terminfo_resolution must be 'dynamic', 'none', or 'static'"
+                            .to_string(),
+                        label: "PythonInterpreterConfig()".to_string(),
                     }));
                 }
             },
-            None => TerminfoResolution::None,
+            None => default.terminfo_resolution,
         };
 
         let sys_paths = match sys_paths.get_type() {
@@ -236,17 +246,18 @@ impl EmbeddedPythonConfig {
             _ => Vec::new(),
         };
 
+        // Automatically enable the filesystem importer if sys.paths are defined.
         let filesystem_importer = filesystem_importer || !sys_paths.is_empty();
 
         Ok(Value::new(EmbeddedPythonConfig {
-            bytes_warning: bytes_warning.to_int().unwrap() as i32,
+            bytes_warning,
             ignore_environment,
             inspect,
             interactive,
             isolated,
             legacy_windows_fs_encoding,
             legacy_windows_stdio,
-            optimize_level: optimize_level.to_int().unwrap(),
+            optimize_level,
             parser_debug,
             quiet,
             stdio_encoding_name,
@@ -262,7 +273,7 @@ impl EmbeddedPythonConfig {
             terminfo_resolution,
             use_hash_seed,
             user_site_directory,
-            verbose: verbose.to_int().unwrap() as i32,
+            verbose,
             write_bytecode,
             write_modules_directory_env,
         }))
@@ -272,41 +283,38 @@ impl EmbeddedPythonConfig {
 starlark_module! { embedded_python_config_module =>
     #[allow(non_snake_case, clippy::ptr_arg)]
     PythonInterpreterConfig(
-        env env,
-        // Keep this in sync with default_starlark() above.
-        bytes_warning=0,
-        ignore_environment=true,
-        inspect=false,
-        interactive=false,
-        isolated=true,
-        legacy_windows_fs_encoding=false,
-        legacy_windows_stdio=false,
-        optimize_level=0,
-        parser_debug=false,
+        bytes_warning=NoneType::None,
+        ignore_environment=NoneType::None,
+        inspect=NoneType::None,
+        interactive=NoneType::None,
+        isolated=NoneType::None,
+        legacy_windows_fs_encoding=NoneType::None,
+        legacy_windows_stdio=NoneType::None,
+        optimize_level=NoneType::None,
+        parser_debug=NoneType::None,
         stdio_encoding=NoneType::None,
-        unbuffered_stdio=false,
-        filesystem_importer=false,
-        quiet=false,
+        unbuffered_stdio=NoneType::None,
+        filesystem_importer=NoneType::None,
+        quiet=NoneType::None,
         run_eval=NoneType::None,
         run_file=NoneType::None,
         run_module=NoneType::None,
-        run_noop=false,
-        run_repl=false,
-        site_import=false,
-        sys_frozen=false,
-        sys_meipass=false,
+        run_noop=NoneType::None,
+        run_repl=NoneType::None,
+        site_import=NoneType::None,
+        sys_frozen=NoneType::None,
+        sys_meipass=NoneType::None,
         sys_paths=NoneType::None,
         raw_allocator=NoneType::None,
-        terminfo_resolution="dynamic",
+        terminfo_resolution=NoneType::None,
         terminfo_dirs=NoneType::None,
-        use_hash_seed=false,
-        user_site_directory=false,
-        verbose=0,
-        write_bytecode=false,
+        use_hash_seed=NoneType::None,
+        user_site_directory=NoneType::None,
+        verbose=NoneType::None,
+        write_bytecode=NoneType::None,
         write_modules_directory_env=NoneType::None
     ) {
         EmbeddedPythonConfig::starlark_new(
-            &env,
             &bytes_warning,
             &ignore_environment,
             &inspect,
@@ -455,6 +463,14 @@ mod tests {
     #[test]
     fn test_terminfo_resolution() {
         let c = starlark_ok("PythonInterpreterConfig(terminfo_resolution=None)");
+        let x = c.downcast_ref::<EmbeddedPythonConfig>().unwrap();
+        assert_eq!(x.terminfo_resolution, TerminfoResolution::Dynamic);
+
+        let c = starlark_ok("PythonInterpreterConfig(terminfo_resolution='dynamic')");
+        let x = c.downcast_ref::<EmbeddedPythonConfig>().unwrap();
+        assert_eq!(x.terminfo_resolution, TerminfoResolution::Dynamic);
+
+        let c = starlark_ok("PythonInterpreterConfig(terminfo_resolution='none')");
         let x = c.downcast_ref::<EmbeddedPythonConfig>().unwrap();
         assert_eq!(x.terminfo_resolution, TerminfoResolution::None);